requirements.txt
django>=5.2
djangorestframework>=3.17
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
psycopg2-binary>=2.9
langchain>=1.2
python-dotenv>=1.0
groq>=1.1
chromadb>=1.5
PyPDF2>=3.0
